- (snmp_agent.c, snmp_api.c, snmp_auth.c, snmp_impl.h):   - fix pdu->versions to be all ints.
git-svn-id: file:///home/hardaker/lib/sf-bkups/net-snmp-convert-svnrepo/trunk@1444 06827809-a52a-0410-b366-d66718629ded

+++ snmplib/snmp_api.c
@@ -813,7 +813,7 @@ snmp_parse(session, pdu, data, length)
     u_char  type;
     struct packet_info pkt, *pi = &pkt;
     u_char  *var_val;
-    long    version;
+    int     version;
     int	    len, four;
     u_char community[COMMUNITY_MAX_LEN];
     int community_length = COMMUNITY_MAX_LEN;
